Struct turbo_tasks_bytes::bytes::Bytes

#[repr(transparent)]
pub struct Bytes(_);

Bytes is a thin wrapper around [bytes::Bytes], implementing easy conversion to/from, ser/de support, and Vc containers.

Implementations

impl Bytes

pub fn cell(self) -> BytesVc

Places a value in a cell of the current task.

Cell is selected based on the value type and call order of cell.

impl Bytes

pub fn to_str(&self) -> Result<&str, Utf8Error>