Struct turbo_tasks_bytes::stream::Stream

pub struct Stream<T: Clone> { /* private fields */ }

Streams allow for streaming values from source to sink.

A Stream implements both a reader (which implements the Stream trait), and a writer (which can be sent to another thread). As new values are written, any pending readers will be woken up to receive the new value.

Implementations

impl<T: Clone> Stream<T>

pub fn new_closed(pulled: Vec<T>) -> Self

Constructs a new Stream, and immediately closes it with only the passed values.

pub fn new_open( pulled: Vec<T>, source: Box<dyn StreamTrait<Item = T> + Send + 'static> ) -> Self

Creates a new Stream, which will lazily pull from the source stream.

pub fn read(&self) -> StreamRead<T>

Returns a [StreamTrait] implementation to poll values out of our Stream.

pub async fn into_single(&self) -> SingleValue<T>

impl<T: Clone, E: Clone> Stream<Result<T, E>>

pub async fn try_into_single(&self) -> Result<SingleValue<T>, E>

Converts a TryStream into a single value when possible.